move get_{platform, distribution, distribution_version} get_all_subclasses and load_platform_subclass into common.sys_info
